+++ src/main/java/store/myproject/onlineshop/domain/membership/dto/MemberShipDto.java
@@ -15,6 +15,6 @@ public class MemberShipDto {
 
     private BigDecimal baseline;
 
-    private Float discountRate;
+    private BigDecimal discountRate;
 
 }
